include process for FreeSurfer to anatomical space
It is just one single command. It does not fit to the scope of fs.py. Again, it cannot fit to the scope of Luigi pipeline. We could write a entire new Python wrapper around it but that would be an overkill given it is just one command. So we are documenting it in our README.md for people to conveniently follow should they need it.

Closes #83

cc @yrathi